Refine server to be deployed in Tomcat
The original version of this server relied on the ability to read
authentication attributes from HTTP request headers. This is a feature
of our authentication system, however it is documented as too risky for
systems and platforms that can use the web server environment. If this
application were put into Tomcat then that becomes an option.

This commit is the original commit of the application inside Tomcat. It
includes:

- The application jar file with no JDK HttpServer code

- Source code developed intended for use with the JDK HttpServer has
  been deleted.

- A bare bones Tomcat 9.0.11 installation with much of the configuration
  removed so that only the items that need configuring are configured.

The application is configured to use the default logging for Tomcat
which is JDK logging in the java.util.logging package. See
src/main/config/logging.properties.

The default servlet and mime-types are the only original configurations
kept for Tomcat. See src/main/config/web.xml.

The server is configured with an HTTP connector for development and an
AJP connector for production. The HTTP connector will be removed prior
to deployment to a production environment. The server is also configured
to use a shared executor pool with a large amount of threads. Many of
the sample resources, valves and listeners have been removed to avoid
security problems. The AJP connector was configured with
tomcatAuthentication set to false to avoid any confusion with the
status of the request after being passed from the web ISO. See
src/main/config/server.xml.

A setenv.sh file was added so that system properties can be set for
elmr. These should be set in the CATALINA_OPTS environment variable.
See src/main/scripts/setenv.sh.

Some webservers (like Apache which we will be working with here) will
squish multivalued attributes into a single line with values delimited
by ';'. LoginServlet was made smart enough to detect this and split
these values up so that the value would render as a list in json.

Packaging was configured so that all the application jars and files are
copied to 'web/elmr' in the deployable artifiact. See
src/main/assembly/distribution.xml.